+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.0.9
+    Fixed Speaker skills
+    Fixed Car Chase
+    Fixed 3d Dice appearing from API rolls logic
+
+
+
+
 v0.0.8
     Redid Decking and Rigging
     Added Spells, Speakers, Rituals and AMPs